feature: inserting middleware to handle errors
