--- src/helpers/bodyParser.js
@@ -1,15 +1,0 @@
-function bodyParser(request, callback) {
-    let body = '';
-
-    request.on('data', (chunk) => {
-        body += chunk;
-    });
-
-    request.on('end', () => {
-        body = JSON.parse(body);
-        request.body = body;
-        callback();
-    });
-}
-
-module.exports = bodyParser;
