+++ src/middlewares/handleErrors.js
@@ -1,0 +1,4 @@
+function handleErrors(error, response, statusCode = 400) {
+    response.writeHead(statusCode, { 'Content-Type': 'application/json' });
+    response.end(JSON.stringify({ error: error.message || "Bad Request" }));
+}
